(feat) heap allocator (WIP)
The heap allocator should be growable and adapt to the std allocator
standard, while also be able to accept our buddy allocator

The current design is to have a heap be a list of subheaps. a subheap
has an allocator (memory.Allocator interface) attached to it and a
memory region (defined by a start ptr and a length) that it manages
allocations on

We also want to adapt the std allocators to our memory.Allocator
interface so that we can (when applicable) create functions to cheaply
check if a subheap can accept an allocation given a size and alignment.
Should be doable for all std allocator we'd want to use (as we're not
using gpa/sbrk/other platform specific allocators)