(fix) packed struct pointers deprecation in zig
On zig master packed struct pointers become bit-level pointers with the
bit we're pointing to encoded in the alignmnent. This means that the
address of the pointer is of type *align(<alignment>, <int size>, <bit
number>) MyStructType instead of *align(1) <fieldType>. This is great
for zig but breaks how we used to use packed structs. We will need to
move off packed struct and replace them all with extern structs +
align(1) for all non ABI aligned fields.