(feat) paging changes (WIP)
The paging system we currently have in place is too complex while also
being too brittle.
In order to make things better, we need to simplify things a lot.
Steps to be taken:
- if we have enough virtual memory we map physical memory (we try to map
all of it, otherwise, boot config should specify a size to map,
otherwise we map an obscene amount like 64TB on PML4)
- the mapping should be simple, we map physical memory at an offset
(page_offset), given by the boot config so that conversion from virt to
phys addresses would be an add
- we need to change the virtual allocator: since all pages are already
allocated, we can greatly simplify how it works maybe work on remaps
- page allocator can be a simple buddy allocator over the already mapped
pages