(fix) rewrite debug subsystem
Zig master changed the interfaces to the debug info/stacktrace
collection
The new API looks for a debug decl in root. In that debug decl the API
expects to see a SelfInfo struct with
init/deinit/getModule/getSymbol/unwindFrame methods that gets used by
the stacktrace collection machinery.
The new API itself is fine from a design POV and the change is
understandable but the default Unwinder for ELF is tightly coupled to
existing OS infrastructure (trying to load elf, etc..) where it could be
more independent I feel like. They also decided to make StackIterator
private for some reason, which feels very arbitrary..
but we move.