(chore) redesign project's structure to allow arch depend extensibility
The current project has a major flaw. it is almost entirely designed
with x64 in mind. although it is highly unlikely that I'd port this to
anything else. I believe it's a good challenge to overcome to design the
kernel to be sufficently modular to permit such port (at least on
paper).
I tried creating common interfaces (through actual interfaces or duck
typing) for submodules but the interplay of "flcn", "arch" and the root
modules makes it impossible (e.g. we could abstract away something
nicely in flcn but then we'll need to call outb/inb which are only
defined in arch). it is clear that we need a dependency between arch and
flcn and between flcn and arch.
The current idea I am exploring is to inverse the problem: what if arch
defines the entrypoint ? we'll have a start function per arch that would
then go on to dictate which subsystems get initialized and when. we
could hide away architecture/platform specific subsystems (apic/smp/vmm)
and keep them tucked away inside their arch modules.
On paper this could solve my problems, we'll see ..